+++ plugins/SitesManager/API.php
@@ -11,6 +11,7 @@ namespace Piwik\Plugins\SitesManager;
 use Exception;
 use Piwik\Access;
 use Piwik\Common;
+use Piwik\Container\StaticContainer;
 use Piwik\Date;
 use Piwik\Db;
 use Piwik\Metrics\Formatter;
@@ -18,10 +19,10 @@ use Piwik\Network\IPUtils;
 use Piwik\Option;
 use Piwik\Piwik;
 use Piwik\ProxyHttp;
+use Piwik\Scheduler\Scheduler;
 use Piwik\SettingsPiwik;
 use Piwik\SettingsServer;
 use Piwik\Site;
-use Piwik\TaskScheduler;
 use Piwik\Tracker;
 use Piwik\Tracker\Cache;
 use Piwik\Url;
@@ -351,7 +352,10 @@ class API extends \Piwik\Plugin\API
      */
     public function getSitesIdWithAtLeastViewAccess($_restrictSitesToLogin = false)
     {
-        if (Piwik::hasUserSuperUserAccess() && !TaskScheduler::isTaskBeingExecuted()) {
+        /** @var Scheduler $scheduler */
+        $scheduler = StaticContainer::getContainer()->get('Piwik\Scheduler\Scheduler');
+
+        if (Piwik::hasUserSuperUserAccess() && !$scheduler->isRunningTask()) {
             return Access::getInstance()->getSitesIdWithAtLeastViewAccess();
         }
 
@@ -360,7 +364,7 @@ class API extends \Piwik\Plugin\API
             // but during scheduled task execution, we sometimes want to restrict sites to
             // a different login than the superuser.
             && (Piwik::hasUserSuperUserAccessOrIsTheUser($_restrictSitesToLogin)
-                || TaskScheduler::isTaskBeingExecuted())
+                || $scheduler->isRunningTask())
         ) {
 
             if (Piwik::hasTheUserSuperUserAccess($_restrictSitesToLogin)) {